+++ fuzz/transform.cc
@@ -37,7 +37,7 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
   tjhandle handle = NULL;
   unsigned char *dstBufs[1] = { NULL };
   size_t dstSizes[1] = { 0 }, maxBufSize;
-  int width = 0, height = 0, jpegSubsamp, dstSubsamp, i;
+  int width = 0, height = 0, jpegSubsamp, i;
   tjtransform transforms[1];
 
   if ((handle = tj3Init(TJINIT_TRANSFORM)) == NULL)
@@ -69,9 +69,9 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
 
   transforms[0].op = TJXOP_NONE;
   transforms[0].options = TJXOPT_PROGRESSIVE | TJXOPT_COPYNONE;
-  dstSizes[0] = maxBufSize = tj3JPEGBufSize(width, height, jpegSubsamp);
-  dstBufs[0] = (unsigned char *)tj3Alloc(dstSizes[0]);
-  if (!dstBufs[0])
+  dstSizes[0] = maxBufSize = tj3TransformBufSize(handle, &transforms[0]);
+  if (dstSizes[0] == 0 ||
+      (dstBufs[0] = (unsigned char *)tj3Alloc(dstSizes[0])) == NULL)
     goto bailout;
 
   tj3Set(handle, TJPARAM_NOREALLOC, 1);
@@ -98,10 +98,9 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
   transforms[0].op = TJXOP_TRANSPOSE;
   transforms[0].options = TJXOPT_GRAY | TJXOPT_CROP | TJXOPT_COPYNONE |
                           TJXOPT_OPTIMIZE;
-  dstSizes[0] = maxBufSize =
-    tj3JPEGBufSize((height + 1) / 2, (width + 1) / 2, TJSAMP_GRAY);
-  dstBufs[0] = (unsigned char *)tj3Alloc(dstSizes[0]);
-  if (!dstBufs[0])
+  dstSizes[0] = maxBufSize = tj3TransformBufSize(handle, &transforms[0]);
+  if (dstSizes[0] == 0 ||
+      (dstBufs[0] = (unsigned char *)tj3Alloc(dstSizes[0])) == NULL)
     goto bailout;
 
   if (tj3Transform(handle, data, size, 1, dstBufs, dstSizes,
@@ -120,14 +119,9 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
 
   transforms[0].op = TJXOP_ROT90;
   transforms[0].options = TJXOPT_TRIM | TJXOPT_ARITHMETIC;
-  dstSubsamp = jpegSubsamp;
-  if (dstSubsamp == TJSAMP_422) dstSubsamp = TJSAMP_440;
-  else if (dstSubsamp == TJSAMP_440) dstSubsamp = TJSAMP_422;
-  else if (dstSubsamp == TJSAMP_411) dstSubsamp = TJSAMP_441;
-  else if (dstSubsamp == TJSAMP_441) dstSubsamp = TJSAMP_411;
-  dstSizes[0] = maxBufSize = tj3JPEGBufSize(height, width, dstSubsamp);
-  dstBufs[0] = (unsigned char *)tj3Alloc(dstSizes[0]);
-  if (!dstBufs[0])
+  dstSizes[0] = maxBufSize = tj3TransformBufSize(handle, &transforms[0]);
+  if (dstSizes[0] == 0 ||
+      (dstBufs[0] = (unsigned char *)tj3Alloc(dstSizes[0])) == NULL)
     goto bailout;
 
   if (tj3Transform(handle, data, size, 1, dstBufs, dstSizes,