+++ java/org_libjpegturbo_turbojpeg_TJTransformer.h
@@ -17,6 +17,14 @@ JNIEXPORT void JNICALL Java_org_libjpegturbo_turbojpeg_TJTransformer_setICCProfi
 
 /*
  * Class:     org_libjpegturbo_turbojpeg_TJTransformer
+ * Method:    bufSize
+ * Signature: (Lorg/libjpegturbo/turbojpeg/TJTransform;)I
+ */
+JNIEXPORT jint JNICALL Java_org_libjpegturbo_turbojpeg_TJTransformer_bufSize
+  (JNIEnv *, jobject, jobject);
+
+/*
+ * Class:     org_libjpegturbo_turbojpeg_TJTransformer
  * Method:    init
  * Signature: ()V
  */